refactor(settings): move run-label accessors, consts, validators to settings_labels.go (PRD #1021 M2)
